+++ changelog.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-Last change: 10-DEC-2025 CEST
+Last change: 11-DEC-2025 CEST
 
 0.9.2
 -----
@@ -22,6 +22,7 @@ Important updates:
 - Simulate: fixed repetead axes tick labels
 - SPINNA: fixed NND plot showing bins/lines outside of xlim
 - SPINNA: extract the picked area based on the last .yaml file entry, not the first one (fixes the issue of incorrect densities extracted for localizations that were picked multiple times)
+- SPINNA: enforce repeated generation of the search space when exp. data/densities/masks change
 
 0.9.0-1
 -------